requirements.txt
streamlit
openai
pandas
pillow
python-dotenv
pdf2image
google-genai
openpyxl

